Remove rx java and replace with coroutines. Add permission handling Much more refactoring.

+++ app/src/main/java/com/gregmcgowan/fivesorganiser/importContacts/Contact.kt
@@ -3,4 +3,4 @@ package com.gregmcgowan.fivesorganiser.importContacts
 data class Contact(val name: String = "",
                    val phoneNumber: String = "",
                    val emailAddress : String = "",
-                   val contactId: Int)
+                   val contactId: Long)
